server/requirements.txt
fastapi
uvicorn[standard]
supabase
python-jose[cryptography]
google-cloud-aiplatform
python-multipart
gunicorn
